pyproject.toml
[tool.poetry]
name = "dibug"
version = "0.2.0"
description = "Debugging Tool for discord.py"
authors = ["Starcea <stardev.uwu@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/star0202/dibug"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = ">=2.0.0"
async-eval = "^0.1.11"
psutil = "^5.9.5"
types-setuptools = "^67.8.0.0"
types-psutil = "^5.9.5.13"

[tool.poetry.group.style.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[tool.poetry.group.type.dependencies]
mypy = "^1.3.0"

[tool.isort]
profile = "black"

[tool.mypy]
files = ["dibug", "typings"]
strict = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
